% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bd4tobd5}
\alias{bd4tobd5}
\title{Convert a format 4 binary dosage file to Format 5}
\usage{
bd4tobd5(bdfile, bdose_file)
}
\arguments{
\item{bdfile}{File name of the format 4 binary dosage file. Format 4 stores
all data in a single file.}

\item{bdose_file}{Path for the output .bdose file. The companion .bdi
metadata file is written to \code{paste0(bdose_file, ".bdi")}.}
}
\value{
NULL (invisibly)
}
\description{
Reads a format 4 binary dosage file (subformat 1, 2, 3, or 4) and converts
it to a Format 5 file pair. If the source file does not contain genotype
probabilities (subformats 1 and 3), those values are stored as missing in
the output.
}
\keyword{internal}
