% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBInaryDosage.R
\name{CreateSubjectDataTable}
\alias{CreateSubjectDataTable}
\title{Function to create data table to contain dosage values}
\usage{
CreateSubjectDataTable(binaryFileInfo)
}
\arguments{
\item{binaryFileInfo}{A list returned from GetBinaryFileInfo}
}
\value{
A data.table with the subject the information
}
\description{
Function to create data table to contain dosage values
This data table is passed to the dosage extraction routines.
The dosage values are then added to this data table
}
