% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindPointer}
\alias{FindPointer}
\title{Function to find the pointer to a column in a data.table}
\usage{
FindPointer(x, y, n)
}
\arguments{
\item{x}{Column in data.table}

\item{y}{Array for storing pointers}

\item{n}{Location in array to store pointer}
}
\value{
0 Success
1 failure
}
\description{
Function to find the pointer to a column in a data.table
}
