% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSNPValues.R
\name{GetAlternateAlleleFrequencies}
\alias{GetAlternateAlleleFrequencies}
\title{Function to calculate alternate allele frequencies}
\usage{
GetAlternateAlleleFrequencies(fileInfo, SNPs, Subjects)
}
\arguments{
\item{fileInfo}{Information about the binary dosage file return from
GetBDoseInfo}

\item{SNPs}{Vector of either SNP names or indices in binary dosage
file to calculate allele frequencies for}

\item{Subjects}{Vector of either Subject IDs of indices in binary dosage
file to calcualte allele for}
}
\value{
List with information about the file including subject and
SNP information
}
\description{
Function to calculate alternate allele frequencies
}
