% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getbdinfo}
\alias{getbdinfo}
\title{Get information about a binary dosage file}
\usage{
getbdinfo(bdfiles)
}
\arguments{
\item{bdfiles}{Vector of file names. For Format 5 files, a single .bdose
file name; the companion .bdi metadata file is read automatically from
\code{paste0(bdfiles[1], ".bdi")}. For format 4, a single file name. For
formats 1, 2, and 3, a vector of three file names: the binary dosage file,
the family information file, and the SNP information file.}
}
\value{
List with information about the binary dosage file.
This includes family and subject IDs along with
a list of the SNPs in the file. Other information needed
to read the file is also included.
}
\description{
Routine to return information about a binary dosage file.
This information is used by other routines to
allow for quicker extraction of values from the
file.
}
\examples{
vcf1abdfile <- system.file("extdata", "vcf1a.bdose", package = "BinaryDosage")
bdinfo <- getbdinfo(bdfiles = vcf1abdfile)
}
