% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBDoseInfo.R
\name{GetBDoseInfo}
\alias{GetBDoseInfo}
\title{Function to read in information about binary dosage file}
\usage{
GetBDoseInfo(bdFile, famFile = "", mapFile = "", index = FALSE,
  sep = "\\t")
}
\arguments{
\item{bdFile}{Name of the binary dosage file}

\item{famFile}{Name of file with subject data. This file is in the format
used by plink .fam files. Not needed for format 4 or greater
binary dosage files.}

\item{mapFile}{Name of file with SNP information. This file is in the format
used by plink for extended map file files. Not needed for
format 4 or greater binary dosage files.}

\item{index}{Index the SNPs for faster reading.}

\item{sep}{Separator in fam and map files}
}
\value{
List with information about the file including subject and
SNP information
}
\description{
Function to read in information about binary dosage file.
This information is used when read the dosage and genetic
probabilities from the file
}
