% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBinaryDosageInfo.R
\name{GetBinaryDosageInfo}
\alias{GetBinaryDosageInfo}
\title{Function to read in information about binary dosage file}
\usage{
GetBinaryDosageInfo(bdFile, famFile = "", mapFile = "")
}
\arguments{
\item{bdFile}{Name of the binary dosage file}

\item{famFile}{Name of file with subject data. This file is in the format
used by plink .fam files. Not needed for format 4 or greater
binary dosage files.}

\item{mapFile}{Name of file with SNP information. This file is in the format
used by plink for extended map file files. Not needed for
format 4 or greater binary dosage files.}
}
\value{
List with information about the file including subject and
SNP information
}
\description{
Function to read in information about binary dosage file.
This information is used when read the dosage and genetic
probabilities from the file
}
