% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBInaryDosage.R
\name{GetBinaryFileInfo}
\alias{GetBinaryFileInfo}
\title{Function to read in information about binary dosage file}
\usage{
GetBinaryFileInfo(filename)
}
\arguments{
\item{filename}{Name of the binary dosage file}
}
\value{
List with information about the file including subject and
SNP information
}
\description{
Function to read in information about binary dosage file.
This information is used when read the dosage and genetic
probabilities from the file
}
