% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDosages}
\alias{GetDosages}
\title{Function to test getting dosage values}
\usage{
GetDosages(bdInfo, snpNumber)
}
\arguments{
\item{bdInfo}{Binary dosage info list}
}
\value{
List of dosages
}
\description{
Function to test getting dosage values
}
