% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSNPValues.R
\name{GetSNPValues}
\alias{GetSNPValues}
\title{Function to read SNP values}
\usage{
GetSNPValues(fileInfo, SNPs, subjects, geneProb = TRUE)
}
\arguments{
\item{fileInfo}{List returned from GetVCFInfo or GetBDoseInfo with value
index set to TRUE}

\item{SNPs}{List of SNPs to read from the file. Can be either a character
or integer vector. Routine is much faster if a sorted integer
vector is provided.}

\item{subjects}{List of subjects to read SNP values for. This can be either
a character or integer vector. If no value is provided SNP
values are read for all subjects}

\item{geneProb}{Indicator if genetic probabilities are to be returned. The
default value is TRUE}
}
\value{
A matrix with the dosage and genetic probabilities in the
columns and each row corresponds to one subject.
}
\description{
Function to read SNP values from VCF or Binary Dosage files.
}
\examples{
# Under construction
}
