% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVCF.R
\name{getvcfinfo}
\alias{getvcfinfo}
\title{Function to return information about a VCF file}
\usage{
getvcfinfo(filename, gzipped = FALSE, index = TRUE, snpidformat = 0L,
  infofile = "")
}
\arguments{
\item{filename}{Name of the VCF file.}

\item{gzipped}{Indicator if VCF file is compressed using gzip.
Default value is FALSE.}

\item{index}{Indicator if file should be indexed. This
allows for faster reading of the file. Indexing a gzipped
file does not speed up access. Default value is TRUE.}

\item{snpidformat}{The format that the SNP ID will be saved as.
0 - same as in the VCF file
1 - <chromosome>:<location>
2 - <chromosome>:<location>:<reference allele>:<alternate allele>
If snpidformat is 1 and the VCF file uses format 2, an error is
generated. Default value is 0.}

\item{infofile}{Name of file name with information about the
imputation of the SNPs. This file is produced by minimac 3 and 4.}
}
\value{
List containing information about the VCF file
to include file name, subject IDs, and information about
the SNPs. Indices for faster reading will be included
if index is set to TRUE
}
\description{
Function to return information about a VCF file.
This information is used by other routines to
allow for quicker extraction of values from the
file.
}
\examples{
# Under construction
}
