% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetVCFInfo_C}
\alias{GetVCFInfo_C}
\title{Function to get information about a VCF file}
\usage{
GetVCFInfo_C(vcfFile)
}
\arguments{
\item{vcfFilename}{Name of VCF file}
}
\value{
List of information about the VCF file
}
\description{
Function to get information about a VCF file.
This information is used to get dosage and genetic
probabilities from the file.
}
