% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFtoBD.R
\name{GetVCFInfoC}
\alias{GetVCFInfoC}
\title{Function to read information about a VCF file}
\usage{
GetVCFInfoC(vcfFile, index = TRUE)
}
\arguments{
\item{vcfFile}{Name of the VCF file}

\item{index}{Indicator if the file should be indexed
for quicker reader. This is not needed if only the
list of subjects and SNPs is needed. The default value
is TRUE.}
}
\value{
List of subjects and SNPs in VCF file. Indices
for reading the file are included if value index = TRUE
was passed to function.
}
\description{
Function reads information about a VCf file. This
information is used to efficiently read dosages and
genetic probabilities from the VCF file using the
GetSNPValues function
}
\examples{
# Under construction
}
