% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Merge42C}
\alias{Merge42C}
\title{Function to merge binary dosage files}
\usage{
Merge42C(mergeFilename, filenames, mergeInfo)
}
\arguments{
\item{mergeFilename}{Name of file containing merged data}

\item{filenames}{Names of files to merge}

\item{mergeInfo}{Information on files to merge and intersection of SNPs
and union of subjects}
}
\value{
0 - success
1 - failure
}
\description{
Function to merge binary dosage files
Files need to be in format 4.2
}
