% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OpenBDFiles}
\alias{OpenBDFiles}
\title{Function to test opening an array of binary dosage files}
\usage{
OpenBDFiles(bdList, mergeFilename, snpLoc)
}
\arguments{
\item{bdList}{List of binary dosage info lists}

\item{mergeFilename}{Name of file to merge data into}

\item{snpLoc}{Locations of snps in the various binary dosage files}
}
\value{
0 Success
1 failure
}
\description{
Function to test opening an array of binary dosage files
}
