% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrintPointer}
\alias{PrintPointer}
\title{Function to print pointers from an array}
\usage{
PrintPointer(y)
}
\arguments{
\item{y}{Array of stored pointers}
}
\value{
0 Success
1 failure
}
\description{
Function to print pointers from an array
}
