% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadBDInfo}
\alias{ReadBDInfo}
\title{Function to convert a VCF file to a binary dosage file for GxEScan}
\usage{
ReadBDInfo(binaryDosageFilename)
}
\arguments{
\item{binaryDosagFilename}{Name of binary dosage file}
}
\value{
0 Success
1 Failure
}
\description{
Function to convert a VCF file to a binary dosage file for GxEScan.
The binary formatted file is smaller that the VCF file and reads in much quicker.
}
