% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadBDSNPs}
\alias{ReadBDSNPs}
\title{Function read binary dosage file SNPs into a data.table}
\usage{
ReadBDSNPs(filename, numSubjects, numSNPs, dosageStart, dataSize, snps,
  dosageptr, p0ptr, p1ptr, p2ptr)
}
\arguments{
\item{filename}{Name of binary dosage file}

\item{numSubjects}{Number of subjects in file}

\item{numSNPs}{Number of SNPs in file}

\item{dosageStart}{Where dosage values start in file}

\item{dataSize}{Size of SNP data on disk}

\item{snps}{List of SNPs to extract}

\item{dosageptr}{Pointers to dosage values in data.table}

\item{p0ptr}{Pointers to p0 values in data.table}

\item{p1ptr}{Pointers to p1 values in data.table}

\item{p2ptr}{Pointers to p2 values in data.table}
}
\value{
0 Success
1 failure
}
\description{
Function read binary dosage file SNPs into a data.table
}
