% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBinaryDosage.R
\name{ReadBinarySNPData}
\alias{ReadBinarySNPData}
\title{Function to read SNPs from a binary dosage file}
\usage{
ReadBinarySNPData(binaryFileInfo, subjectData, SNPNumbers, maxSNPs)
}
\arguments{
\item{binaryFileInfo}{A list returned from GetBinaryFileInfo}

\item{subjectData}{Data table with subject data - returned from CreateSubjectDataTable}

\item{SNPNumbers}{Integer vector on SNPs to retrieve}

\item{maxSNPs}{Maximum number of SNPs to read in}
}
\value{
List with information about what SNPs were read in
This list may be pass to ReadMoreSNPs to retrieve additional
SNPs if the number of SNPs specified in SNPNumbers is larger
that maxSNPs
}
\description{
Function to read SNPs from a binary dosage file
GetBinaryFileInfo and CreateSubjectDataTable need
to be run first
}
