% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBInaryDosage.R
\name{ReadMoreBinarySNPData}
\alias{ReadMoreBinarySNPData}
\title{Function to read SNPs from a binary dosage file}
\usage{
ReadMoreBinarySNPData(binaryFileInfo, readInfo)
}
\arguments{
\item{binaryFileInfo}{Information about binary file returned from GetBinaryFileInfo}

\item{readInfo}{A list returned from ReadBinarySNPData}
}
\value{
List with information about what SNPs were read in
This list may be pass to ReadMoreBinarySNPData to retrieve additional
SNPs if the number of SNPs specified in SNPNumbers is larger
that maxSNPs
}
\description{
Function to read SNPs from a binary dosage file
GetBinaryFileInfo and CreateSubjectDataTable need
to be run first
}
