% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBInaryDosage.R
\name{SNPApply}
\alias{SNPApply}
\title{Function to loop through the SNPs in a binary dosage file}
\usage{
SNPApply(binaryDosageInfo, function2apply, outputValues,
  extraData = NULL, snps2use, maxSNPs)
}
\arguments{
\item{binaryDosageInfo}{Information returned from GetBinaryFileInfo}

\item{function2apply}{Function to be preformed on each SNP}

\item{outputValues}{Place to store the results of the function
This is usually a data.table because it can be modified by
function}

\item{extraData}{Addition data to be passed to the function to apply}

\item{snps2use}{List of indices of the SNPs in the binary dosage file to use}

\item{maxSNPs}{Maximum number of SNPs to read in at one time - current limit is 200
This needs work. It is an issue with data.tables}
}
\value{
0 - success - currently always succeeds
1 - failure
The results of the function to apply are stored in outputValues
}
\description{
Function to loop through the SNPs in a binary dosage file
and apply a function to each each SNP
}
