% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestWriteBinaryDosage}
\alias{TestWriteBinaryDosage}
\title{Function to test writing of binary dosage files}
\usage{
TestWriteBinaryDosage()
}
\value{
0 success
1 failure
}
\description{
Function to test writing of binary dosage files
}
