% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VCF2BD}
\alias{VCF2BD}
\title{Function to convert a VCF file to a binary dosage file for GxEScan}
\usage{
VCF2BD(vcfFilename, infoFilename, outputFilename)
}
\arguments{
\item{vcfFilename}{Name of VCF file}

\item{infoFilename}{Name of information file}

\item{outputFilename}{Name of the binary dosage file}
}
\value{
0 Success
1 Failure
}
\description{
Function to convert a VCF file to a binary dosage file for GxEScan.
The binary formatted file is smaller that the VCF file and reads in much quicker.
}
