% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VCF2Impute2BD}
\alias{VCF2Impute2BD}
\title{Function to convert an Impute 2 dosage file to a binary dosage file for GxEScan}
\usage{
VCF2Impute2BD(vcfFilename, infoFilename, sampleFilename, outputFilename)
}
\arguments{
\item{vcfFilename}{Name of VCF file}

\item{infoFilename}{Name of information file}

\item{outputFilename}{Name of the binary dosage file}

\item{smapleFilename}{Name of sample file - file with subject IDs}
}
\value{
0 Success
1 Failure
}
\description{
Function to convert a VCF file to a binary dosage file for GxEScan.
The Impute 2 dosage file has to be created from a VCF file using VCF tools
The binary formatted file is smaller that the VCF file and reads in much quicker.
}
