% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVCF.R
\name{vcfapply}
\alias{vcfapply}
\title{Apply a function to all the SNPs in a VCF file}
\usage{
vcfapply(vcfinfo, func, ...)
}
\arguments{
\item{vcfinfo}{Information about a VCF file returned from
getvcfinfo}

\item{func}{The function to apply to each of the SNPs.}
}
\value{
A list with the function result for each SNP
}
\description{
Routine to apply a user provided function to all the
SNPs in a VCF file.
}
\examples{
# In work
}
