% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VCFBBtoBD_C}
\alias{VCFBBtoBD_C}
\title{Function to convert a BioBank VCF file from HRC to a binary dosage file for GxEScan without an info file}
\usage{
VCFBBtoBD_C(vcfFilename, bdFilename)
}
\arguments{
\item{vcfFilename}{Name of VCF file}

\item{outputFilename}{Name of the binary dosage file}
}
\value{
0 Success
1 Failure
}
\description{
Function to convert a VCF file from HRC to a binary dosage file for GxEScan without an info file.
The binary formatted file is smaller that the VCF file and reads in much quicker.
}
