% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBinaryDosage.R
\name{getbdinfo}
\alias{getbdinfo}
\title{Function to read information about binary dosage files}
\usage{
getbdinfo(bdfilenames)
}
\arguments{
\item{bdfilenames}{- Vector of file names. The first is the
binary dosage data containing the dosages and genetic
probabilites. The second file name is the family information
file. The thrid file name is the SNP information file.
The family and SNP information files are not used if the
binary dosage file is in format 4. For this format the
family and SNP information are in the file with the dosage
and genetic probabilites.}
}
\value{
List with subject and SNP information and values needed
to read the dosages and genetic probabilities,
}
\description{
Function to read information about binary dosage files.
This information is used when reading the dosage and genetic
probabilities from the file. It is also used by GxEScanR.
}
\examples{
# in work
}
