% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSNPValues.R
\name{GetSNPValues}
\alias{GetSNPValues}
\title{Function to read in SNPs from a binary dosage file}
\usage{
GetSNPValues(bdInfo, SNPs, Subjects)
}
\arguments{
\item{bdInfo}{Information about the binary dosage file return from
GetBDoseInfo}

\item{SNPs}{Vector of either SNP names or indices in binary dosage
file to extract values for}

\item{Subjects}{Vector of either Subject IDs of indices in binary dosage
file to extract values for}
}
\value{
List with information about the file including subject and
SNP information
}
\description{
Function to read in SNPs from a binary dosage file
}
