% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetSNPValues}
\alias{GetSNPValues}
\title{Function to get infromation about a binary dosage file}
\usage{
GetSNPValues(bdFilename, n)
}
\arguments{
\item{bdFilename}{Name of binary dosage file}

\item{famFilename}{Name of subject data file}

\item{mapFilename}{Name of SNP data file}
}
\value{
List with information about the binary dosage file
}
\description{
Function to get infromation about a binary dosage file
}
