% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFtoBD.R
\name{GetVCFInfo}
\alias{GetVCFInfo}
\title{Function to return information about a VCF file}
\usage{
GetVCFInfo(filename, gzipped = FALSE, index = TRUE)
}
\arguments{
\item{filename}{Name of the VCF file.}

\item{gzipped}{Indicator if VCF file is compressed using gzip.
Default value is FALSE.}

\item{index}{Indicator if file should be indexed. This
allows for faster reading of the file. Indexing a gzipped
file does not speed up access. Default value is TRUE.}
}
\value{
List containing information about the VCF file
to include file name, subject IDs, and information about
the SNPs. Indices for faster reading will be included
if index is set to TRUE
}
\description{
Function to return information about a VCF file.
This information is used by other routines to
allow for quicker extraction of values from the
file.
}
\examples{
# Under construction
}
