% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVCFInfo.R
\name{GetVCFInfo}
\alias{GetVCFInfo}
\title{Function to read in information about a vcf file}
\usage{
GetVCFInfo(vcfFile, reserve)
}
\arguments{
\item{vcfFile}{Name of the vcf file}
}
\value{
List with information about the file including subject and
SNP information
}
\description{
Function to read in information about a vcf file.
This information is used when reading the dosage and genetic
probabilities from the file
}
