% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Merge42C}
\alias{Merge42C}
\title{Function to test getting dosage values}
\usage{
Merge42C(mergeFilename, filenames, mergeInfo)
}
\arguments{
\item{mergeFilename}{Name of file containing merged data}

\item{filenames}{Names of files to merge}

\item{mergeInfo}{Information on files to merge and intersection of SNPs
and union of subjects}
}
\value{
List of dosages
}
\description{
Function to test getting dosage values
}
