% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{MergeBD}
\alias{MergeBD}
\title{Function to merge binary dosage files in format 4.2}
\usage{
MergeBD(mergedFile, filesToMerge, famFilesToMerge, mapFilesToMerge,
  mergedFamFile = "", mergeMapFile = "", format = 4, version = 2)
}
\arguments{
\item{filesToMerge}{A vector of strings. Names of the files to merge.}

\item{mergeFile}{Name of file that will contain the merged data.}
}
\value{
0 - Successfully merged
1 - Merge failed
}
\description{
Function to merge binary dosage files in format 4.2
}
