% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{MergeBD}
\alias{MergeBD}
\title{Function to merge binary dosage files}
\usage{
MergeBD(mergedFile, filesToMerge, famFilesToMerge, mapFilesToMerge,
  mergedFamFile = "", mergeMapFile = "", format = 4, version = 2,
  batchSize = 100)
}
\arguments{
\item{filesToMerge}{A vector of strings. Names of the files to merge.}

\item{famFilesToMerge}{A vector of strings. These are the names of the fam files
associated with files to merge. If the files to merge are
in format 4 or later, this is not needed.}

\item{mapFilesToMerge}{A vector of strings. These are the names of the map files
associated with files to merge. If the files to merge are
in format 4 or later, this is not needed.}

\item{mergedFamFile}{Name of fam file to be associated with the binary dosage
file. This is only needed if format is 3 or less.}

\item{format}{Format of the merged binary dosage file. Default value is 4.
It is recommended to use the default value.}

\item{version}{Version of the format to use. Default value is 2. If the files
being merged do not contain gene probabilities, change this
value to 1.}

\item{batchSize}{Number of SNPs to process at a time}

\item{mergeFile}{Name of file that will contain the merged data.}

\item{mergedMapFile}{Name of map file to be associated with the binary dosage
file. This is only needed if format is 3 or less.}
}
\value{
0 - Successfully merged
1 - Merge failed
}
\description{
Function to merge binary dosage files
}
