% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF2BD.R
\name{VCF2BD}
\alias{VCF2BD}
\title{Function to creae a binary dosage file from a VCF file}
\usage{
VCF2BD(VCFFilename, outputFilename, infoFilename = "",
  calculateMAF = TRUE)
}
\arguments{
\item{VCFFilename}{VCF file name}

\item{outputFilename}{Name of output file}

\item{infoFilename}{Name of info file associated with VCF file. Can be blank.
Default value = ""}

\item{calculateMAF}{Indicator on whether to calculate MAF based on the sample.
Default value = TRUE}
}
\value{
0 - success
1 - failure
}
\description{
Function to creae a binary dosage file from a VCF file
}
