% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFtoBD.R
\name{vcftobd}
\alias{vcftobd}
\title{Function to convert a VCF file to a binary dosage file}
\usage{
vcftobd(vcffile, vcfinfofile = "", gz = FALSE, bdfiles, format = 4L,
  subformat = 0L, snpidformat = 0, bdoptions = character(0))
}
\arguments{
\item{vcffile}{Name of VCF file}

\item{vcfinfofile}{(Optional) Name of information file associated
with the vcf file. Default value "".}

\item{gz}{Indicator if vcf file in compressed using gzip.
The default value is FALSE.}

\item{bdfiles}{Vector of names of the output files.
The binary dosage file name is first. The family and
map files follow. For format 4, no family and map file
names are needed.}

\item{format}{The format of the output binary dosage file.
Allowed values are 1, 2, 3, and 4. The default value is 4.
Using the default value is recommended.}

\item{subformat}{The subsubformat of the format of the output
binary dosage file. A value of 1 or 3 indicates that only the
dosage value is saved. A value of 2 or 4 indicates
the dosage and genetic probabilities will be output. Values
of 3 or 4 are only allowed with formats 3 and 4. If a value
of zero if provided, and genetic probabilites are in the vcf
file, subformat 2 will be used for formats 1 and 2, and
subformat 4 will be used for formats 3 and 4. If the vcf file
does not contain genetic probabilities, subformat 1 will be
used for formats 1 and 2, and subformat 3 will be used for
formats 3 and 4. The default value is 0.}

\item{snpidformat}{Format to store the snp id in. Format 1
is chromosome:location. Format 2 is
chromosome:location:reference:alternate. Format 0 indicates
to use the format in the vcf file. 0 is the default.}

\item{bdoptions}{Character array containg any of the following
value, "aaf", "maf", "rsq". The presence of any of these
values indicates that the specified values should be
calculates and stored in the bdosage file. These values only
apply to format 4.}
}
\value{
A list containing information about the binary dosage file.
This is the same list returned from BDInfo. See
GetBDoseInfo for more information.
}
\description{
Function to read information from a VCF file and create
a binary dosage file. The function is designed to use
files return from the Michigan Imputation Server but will
run on other VCF files if they contain dosage and genetic
probabilities.
}
\examples{
# Under construnction
}
