% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBD.R
\name{VCFtoBD}
\alias{VCFtoBD}
\title{Function to read in information about a vcf file}
\usage{
VCFtoBD(vcfFile, bdFile, famFile = "", mapFile = "", format = 4,
  version = 2, batchSize = 100)
}
\arguments{
\item{vcfFile}{Name of the vcf file}

\item{bdFile}{Name of the binary dosage file}

\item{famFile}{Name of the family file, only used with formats 1, 2, and 3}

\item{mapFile}{Name of the map file, only used with formats 1, 2, and 3}

\item{format}{Format of binary dosage file - default 4, can be 1, 2, or 3}

\item{version}{Version of the format to use - default 2, can be 1}

\item{batchSize}{Size of batches of SNPs to read in before processing}
}
\value{
0 - Success
1 - Failure
}
\description{
Function to read in information about a vcf file.
This information is used when reading the dosage and genetic
probabilities from the file
}
