% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFtoBD.R
\name{VCFtoBD}
\alias{VCFtoBD}
\title{Function to convert a VCF file to a binary dosage file}
\usage{
VCFtoBD(vcfFile, bdFile, famFile = "", mapFile = "", gz = FALSE,
  format = 4L, version = 0L)
}
\arguments{
\item{vcfFile}{Name of VCF file}

\item{bdFile}{Name of the output binary dosage file}

\item{famFile}{Name of output family information file.
This is only used if the output format is 1, 2, or 3.
The default value is ""}

\item{mapFile}{Name of the output map information file.
This is only used if the output format is 1, 2, or 3.
The default value is ""}

\item{gz}{Indicatator if vcf file is compressed using gzip.}

\item{format}{The format of the output binary dosage file.
Allowed values are 1, 2, 3, and 4. The default value is 4.
Using the default value is recommended.}

\item{version}{The subversion of the format of the output
binary dosage file. A value of 1 indicates that only the
dosage value is saved. A value greater than 1 indicates
the dosage and genetic probabilities will be output. A
value of 0 indicates the genetic probabilities will be
output if they are contained in the VCF file. Values of
1 and 2 are currently supported for all formats.
The default value is 0.}
}
\value{
A list containing information about the binary dosage file.
This is the same list returned from GetBDoseInfo. See
GetBDoseInfo for more information.
}
\description{
Function to read information from a VCF file and create
a binary dosage file. The function is designed to use
files return from the Michigan Imputation Server but will
run on other VCF files if they contain dosage and genetic
probabilities.
}
\examples{
# Under construnction
}
