% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{vcftobd5}
\alias{vcftobd5}
\title{Convert VCF file to a Binary Dosage file format 5}
\usage{
vcftobd5(vcffile, bdfile, bdinfo, dosageonly = FALSE)
}
\arguments{
\item{vcffile}{Name of VCF file}

\item{bdfile}{Name of Binary Dosage file}

\item{bdinfo}{Name of Binary Dosage information file}

\item{dosageOnly}{Indicator if only dosage values should be saved.
default:FALSE}
}
\value{
List containing Binary Dosage information on success, Empty List on failure.
}
\description{
Convert VCF file to a Binary Dosage file format 5
}
\examples{
vcftobd5("chr22.vcf", "chr22.bdose", "chr22.bdinfo")
}
